+++ proto/sdk/builder/v1/genesis.proto
@@ -6,7 +6,7 @@ import "cosmos/base/v1beta1/coin.proto";
 import "amino/amino.proto";
 import "cosmos_proto/cosmos.proto";
 
-option go_package = "github.com/skip-mev/block-sdk/x/builder/types";
+option go_package = "github.com/skip-mev/block-sdk/x/auction/types";
 
 // GenesisState defines the genesis state of the x/builder module.
 message GenesisState { Params params = 1 [ (gogoproto.nullable) = false ]; }
