+++ x/auction/keeper/auction.go
@@ -5,7 +5,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/skip-mev/block-sdk/x/builder/types"
+	"github.com/skip-mev/block-sdk/x/auction/types"
 )
 
 // ValidateBidInfo validates that the bid can be included in the auction.
