+++ x/auction/keeper/msg_server.go
@@ -9,7 +9,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/skip-mev/block-sdk/x/builder/types"
+	"github.com/skip-mev/block-sdk/x/auction/types"
 )
 
 var _ types.MsgServer = MsgServer{}
