+++ x/auction/rewards/fixed_provider.go
@@ -5,7 +5,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/skip-mev/block-sdk/x/builder/types"
+	"github.com/skip-mev/block-sdk/x/auction/types"
 )
 
 var _ types.RewardsAddressProvider = (*FixedAddressRewardsAddressProvider)(nil)
