+++ x/auction/rewards/proposer_provider.go
@@ -3,7 +3,7 @@ package rewards
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/skip-mev/block-sdk/x/builder/types"
+	"github.com/skip-mev/block-sdk/x/auction/types"
 )
 
 var _ types.RewardsAddressProvider = (*ProposerRewardsAddressProvider)(nil)
