+++ x/auction/types/bid_info.go
